Control image scale et déplacement

+++ WorldGeneration/view/MapView.h
@@ -9,12 +9,12 @@ class MapView
 private:
 	MapScreen* mapScreen;
 	std::chrono::time_point<std::chrono::system_clock> percentLastChange = std::chrono::system_clock::now() ;
-	const static unsigned int timeElapsedPercent{ 50 };
+	const static unsigned short timeElapsedPercent{ 50 };
 public:
 	MapView(MapScreen* map_screen) :mapScreen(map_screen) {}
 	void loadMap(const Map& map);
 	void setMessageId(std::string key);
-	void setPercent(int percent);
+	void setPercent(unsigned int percent);
 
 };
